fix build error

+++ src/cpp/math/method/prime/millerrabin.hpp
@@ -15,7 +15,7 @@
 //
 // %require
 // ```
-using ll = int64_t;
+using ll = long long;
 // ```
 //
 // %verified
